requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0.1
albumentations>=1.4.0
opencv-python>=4.9.0
Pillow>=10.0.0